Refactor DfsBlockCache and make it pluggable
- Made a DfsBlockCache abstract class
- Renamed the original DfsBlockCache into DefaultDfsBlockCache
  and refactor to make it extend DfsBlockCache
